Replacing usage of std::size_t with std::int64_t.

+++ Syntropy/include/syntropy/containers/stream_vector.h
@@ -48,32 +48,32 @@ namespace syntropy
 
         /// \brief Returns a reference to the element at specified index in a given stream.
         /// \return Returns the reference to the requested element in the specified stream.
-        template <std::size_t kStream>
-        auto& GetElementAt(std::size_t index);
+        template <std::int64_t kStream>
+        auto& GetElementAt(std::int64_t index);
 
         /// \brief Returns a reference to the element at specified index in a given stream.
         /// \return Returns the reference to the requested element in the specified stream.
-        template <std::size_t kStream>
-        const auto& GetElementAt(std::size_t index) const;
+        template <std::int64_t kStream>
+        const auto& GetElementAt(std::int64_t index) const;
 
         /// \brief Returns a reference to the element at specified index in a given stream type.
         /// \return Returns the reference to the requested element in the specified stream.
         template <typename TElement>
-        TElement& GetElementAt(std::size_t index);
+        TElement& GetElementAt(std::int64_t index);
 
         /// \brief Returns a reference to the element at specified index in a given stream type.
         /// \return Returns the reference to the requested element in the specified stream.
         template <typename TElement>
-        const TElement& GetElementAt(std::size_t index) const;
+        const TElement& GetElementAt(std::int64_t index) const;
 
         /// \brief Returns a reference to the first element of a given stream.
         /// \return Returns the reference to the first element in the specified stream.
-        template <std::size_t kStream>
+        template <std::int64_t kStream>
         auto& GetFront();
 
         /// \brief Returns a reference to the first element of a given stream.
         /// \return Returns the reference to the first element in the specified stream.
-        template <std::size_t kStream>
+        template <std::int64_t kStream>
         const auto& GetFront() const;
 
         /// \brief Returns a reference to the first element of a given stream, by stream type.
@@ -88,12 +88,12 @@ namespace syntropy
 
         /// \brief Returns a reference to the last element of a given stream.
         /// \return Returns the reference to the last element in the specified stream.
-        template <std::size_t kStream>
+        template <std::int64_t kStream>
         auto& GetBack();
 
         /// \brief Returns a reference to the last element of a given stream.
         /// \return Returns the reference to the last element in the specified stream.
-        template <std::size_t kStream>
+        template <std::int64_t kStream>
         const auto& GetBack() const;
 
         /// \brief Returns a reference to the last element of a given stream, by stream type.
@@ -112,13 +112,13 @@ namespace syntropy
 
         /// \brief Get the number of elements in the container.
         /// \return Returns the number of elements in the container.
-        std::size_t GetSize() const noexcept;
+        std::int64_t GetSize() const noexcept;
 
         /// \brief Increase the capacity of the vector to a value that's greater or equal to capacity. If capacity is greater than the current GetCapacity(), new storage is allocated, otherwise the method does nothing.
-        void Reserve(std::size_t capacity);
+        void Reserve(std::int64_t capacity);
 
         /// \brief Get the number of elements that can be contained in the current allocation storage.
-        std::size_t GetCapacity() const noexcept;
+        std::int64_t GetCapacity() const noexcept;
 
         /// \brief Reduce memory usage by reducing unused memory.
         void ShrinkToFit();
@@ -136,14 +136,14 @@ namespace syntropy
         /// \brief Remove the index-th element of the container and replaces it with the last element in the container.
         /// The order of the container is not preserved as a result of this call.
         /// /// \param index Index of the element to remove.
-        void EraseSwap(std::size_t index);
+        void EraseSwap(std::int64_t index);
 
         /// \brief Execute an operation on each element of the given streams.
-        template <std::size_t... kStreams, typename TOperation>
+        template <std::int64_t... kStreams, typename TOperation>
         void ForEach(TOperation&& operation);
 
         /// \brief Execute an operation on each element of the given streams.
-        template <std::size_t... kStreams, typename TOperation>
+        template <std::int64_t... kStreams, typename TOperation>
         void ForEach(TOperation&& operation) const;
 
         /// \brief Execute an operation on each element of the given streams, by stream type.
@@ -155,15 +155,15 @@ namespace syntropy
         void ForEach(TOperation&& operation) const;
 
         /// \brief Get a view into a stream by index.
-        template <std::size_t kStream>
+        template <std::int64_t kStream>
         auto GetStream();
 
         /// \brief Get a view into a constant stream by index.
-        template <std::size_t kStream>
+        template <std::int64_t kStream>
         auto GetStream() const;
 
         /// \brief Get a view into a constant stream by index.
-        template <std::size_t kStream>
+        template <std::int64_t kStream>
         auto GetConstStream() const;
 
         /// \brief Get a view into a stream by type.
@@ -181,11 +181,11 @@ namespace syntropy
     private:
 
         /// \brief Appends the given element values to the end of a given stream.
-        template <std::size_t kStream, typename TValue, typename... TValues>
+        template <std::int64_t kStream, typename TValue, typename... TValues>
         void PushBack(TValue&& value, TValues&&... values);
 
         /// \brief Appends the given element value to the end of a given stream.
-        template <std::size_t kStream>
+        template <std::int64_t kStream>
         void PushBack();
 
         /// \brief Execute an operation on each stream.
@@ -193,7 +193,7 @@ namespace syntropy
         void ForEachStream(TOperation&& operation);
 
         /// \brief Execute an operation on each indexed stream.
-        template <typename TOperation, std::size_t kStreamIndex, std::size_t... kStreamIndexes>
+        template <typename TOperation, std::int64_t kStreamIndex, std::int64_t... kStreamIndexes>
         void ForEachStream(TOperation&& operation, std::index_sequence<kStreamIndex, kStreamIndexes...>);
 
         /// \brief Execute an operation on each indexed stream.
@@ -212,8 +212,8 @@ namespace syntropy
     // StreamVector<T...>.
 
     template <typename... TStreams>
-    template <std::size_t kStream>
-    inline auto& StreamVector<TStreams...>::GetElementAt(std::size_t index)
+    template <std::int64_t kStream>
+    inline auto& StreamVector<TStreams...>::GetElementAt(std::int64_t index)
     {
         using TElement = std::tuple_element_t<kStream, std::tuple<TStreams...>>;
 
@@ -221,28 +221,28 @@ namespace syntropy
     }
 
     template <typename... TStreams>
-    template <std::size_t kStream>
-    inline const auto& StreamVector<TStreams...>::GetElementAt(std::size_t index) const
+    template <std::int64_t kStream>
+    inline const auto& StreamVector<TStreams...>::GetElementAt(std::int64_t index) const
     {
         return std::get<kStream>(streams_)[index];
     }
 
     template <typename... TStreams>
     template <typename TElement>
-    inline TElement& StreamVector<TStreams...>::GetElementAt(std::size_t index)
+    inline TElement& StreamVector<TStreams...>::GetElementAt(std::int64_t index)
     {
         return GetElementAt<tuple_element_index_v<TElement, std::tuple<TStreams...>>>(index);
     }
 
     template <typename... TStreams>
     template <typename TElement>
-    inline const TElement& StreamVector<TStreams...>::GetElementAt(std::size_t index) const
+    inline const TElement& StreamVector<TStreams...>::GetElementAt(std::int64_t index) const
     {
         return GetElementAt<tuple_element_index_v<TElement, std::tuple<TStreams...>>>(index);
     }
 
     template <typename... TStreams>
-    template <std::size_t kStream>
+    template <std::int64_t kStream>
     inline auto& StreamVector<TStreams...>::GetFront()
     {
         using TElement = std::tuple_element_t<kStream, std::tuple<TStreams...>>;
@@ -251,7 +251,7 @@ namespace syntropy
     }
 
     template <typename... TStreams>
-    template <std::size_t kStream>
+    template <std::int64_t kStream>
     inline const auto& StreamVector<TStreams...>::GetFront() const
     {
         return std::get<kStream>(streams_).front();
@@ -272,7 +272,7 @@ namespace syntropy
     }
 
     template <typename... TStreams>
-    template <std::size_t kStream>
+    template <std::int64_t kStream>
     inline auto& StreamVector<TStreams...>::GetBack()
     {
         using TElement = std::tuple_element_t<kStream, std::tuple<TStreams...>>;
@@ -281,7 +281,7 @@ namespace syntropy
     }
 
     template <typename... TStreams>
-    template <std::size_t kStream>
+    template <std::int64_t kStream>
     inline const auto& StreamVector<TStreams...>::GetBack() const
     {
         return std::get<kStream>(streams_).back();
@@ -308,13 +308,13 @@ namespace syntropy
     }
 
     template <typename... TStreams>
-    inline std::size_t StreamVector<TStreams...>::GetSize() const noexcept
+    inline std::int64_t StreamVector<TStreams...>::GetSize() const noexcept
     {
         return std::get<0>(streams_).size();
     }
 
     template <typename... TStreams>
-    inline void StreamVector<TStreams...>::Reserve(std::size_t capacity)
+    inline void StreamVector<TStreams...>::Reserve(std::int64_t capacity)
     {
         ForEachStream([capacity](auto& stream)
         {
@@ -323,7 +323,7 @@ namespace syntropy
     }
 
     template <typename... TStreams>
-    inline std::size_t StreamVector<TStreams...>::GetCapacity() const noexcept
+    inline std::int64_t StreamVector<TStreams...>::GetCapacity() const noexcept
     {
         return std::get<0>(streams_).capacity();
     }
@@ -365,7 +365,7 @@ namespace syntropy
     }
 
     template <typename... TStreams>
-    inline void StreamVector<TStreams...>::EraseSwap(std::size_t index)
+    inline void StreamVector<TStreams...>::EraseSwap(std::int64_t index)
     {
         ForEachStream([index](auto& stream)
         {
@@ -379,7 +379,7 @@ namespace syntropy
     }
 
     template <typename... TStreams>
-    template <std::size_t... kStreams, typename TOperation>
+    template <std::int64_t... kStreams, typename TOperation>
     inline void StreamVector<TStreams...>::ForEach(TOperation&& operation)
     {
         for (auto index = 0u; index < GetSize(); ++index)
@@ -389,7 +389,7 @@ namespace syntropy
     }
 
     template <typename... TStreams>
-    template <std::size_t... kStreams, typename TOperation>
+    template <std::int64_t... kStreams, typename TOperation>
     inline void StreamVector<TStreams...>::ForEach(TOperation&& operation) const
     {
         for (auto index = 0u; index < GetSize(); ++index)
@@ -413,7 +413,7 @@ namespace syntropy
     }
 
     template <typename... TStreams>
-    template <std::size_t kStream>
+    template <std::int64_t kStream>
     inline auto StreamVector<TStreams...>::GetStream()
     {
         using TElement = std::tuple_element_t<kStream, std::tuple<TStreams...>>;
@@ -422,14 +422,14 @@ namespace syntropy
     }
 
     template <typename... TStreams>
-    template <std::size_t kStream>
+    template <std::int64_t kStream>
     inline auto StreamVector<TStreams...>::GetStream() const
     {
         return GetConstStream<kStream>();
     }
 
     template <typename... TStreams>
-    template <std::size_t kStream>
+    template <std::int64_t kStream>
     inline auto StreamVector<TStreams...>::GetConstStream() const
     {
         return MakeConstVectorView(std::get<kStream>(streams_));
@@ -457,7 +457,7 @@ namespace syntropy
     }
 
     template <typename... TStreams>
-    template <std::size_t kStream, typename TValue, typename... TValues>
+    template <std::int64_t kStream, typename TValue, typename... TValues>
     inline void StreamVector<TStreams...>::PushBack(TValue&& value, TValues&&... values)
     {
         std::get<kStream>(streams_).emplace_back(std::forward<TValue>(value));
@@ -466,7 +466,7 @@ namespace syntropy
     }
 
     template <typename... TStreams>
-    template <std::size_t kStream>
+    template <std::int64_t kStream>
     inline void StreamVector<TStreams...>::PushBack()
     {
 
@@ -480,7 +480,7 @@ namespace syntropy
     }
 
     template <typename... TStreams>
-    template <typename TOperation, std::size_t kStreamIndex, std::size_t... kStreamIndexes>
+    template <typename TOperation, std::int64_t kStreamIndex, std::int64_t... kStreamIndexes>
     inline void StreamVector<TStreams...>::ForEachStream(TOperation&& operation, std::index_sequence<kStreamIndex, kStreamIndexes...>)
     {
         operation(std::get<kStreamIndex>(streams_));
